+++ src/main/kotlin/app/morphe/cli/command/MainCommand.kt
@@ -5,11 +5,13 @@ import app.morphe.library.logging.Logger
 import picocli.CommandLine
 import picocli.CommandLine.Command
 import picocli.CommandLine.IVersionProvider
-import java.util.*
+import java.util.Properties
+import kotlin.system.exitProcess
 
 fun main(args: Array<String>) {
     Logger.setDefault()
-    CommandLine(MainCommand).execute(*args).let(System::exit)
+    val exitCode = CommandLine(MainCommand).execute(*args)
+    exitProcess(exitCode)
 }
 
 private object CLIVersionProvider : IVersionProvider {
@@ -36,7 +38,8 @@ private object CLIVersionProvider : IVersionProvider {
         PatchCommand::class,
         ListPatchesCommand::class,
         ListCompatibleVersions::class,
+        OptionsCommand::class,
         UtilityCommand::class,
-    ],
+    ]
 )
 private object MainCommand